pyproject.toml
[project]
name = "ddt-sharp-shooter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddddocr>=1.5.6",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "pyautogui>=0.9.54",
    "pyinstaller>=6.13.0",
    "pynput>=1.8.1",
    "pytesseract>=0.3.13",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
]
